10x 2x fixed everything 2x

+++ src/memory_controller.h
@@ -476,7 +476,7 @@ void reset_profile_counters();
 void calc_wait_value(char op_type, int reqid, long long int comptime, int oramid);
 long long int sum_wait_sofar();
 void set_reshuffle_w(int level);
-void adjust_ddr(int addr);
+void adjust_ddr(long long int addr);
 
 
 // Mehrnoosh.
